Incorporate resource type in UID.
A bridge's fingerprint is ambiguous and could refer to both an obfs3 or
obfs4 endpoint.  To eliminate this ambiguity, this patch incorporates
the resource type into the UID.  Note that this assumes that a bridge
cannot run more than one instance of a given pluggable transport.  This
is true for now but may not be true in the future.