Replace incremental IDs with secret passwords.
So far, we used incremental user IDs as secret identifiers.  This patch
replaces those user IDs with secret passphrases that serve both as user
IDs and as a way for users to authenticate themselves to the backend.
Secret user IDs consist of 20 random bytes encoded in Base32.

This fixes tpo/anti-censorship/rdsys#15