Use separate struct to map proxies to users.
Having pointers in the User and Proxy struct is confusing and difficult
to maintain.  This patch implements a separate data structure,
ProxyAssignments, that keeps track who got what proxy.

We currently use ProxyAssignments only in Salmon but we can generalise
the concept and move it to the domain layer if we ever need it in a
different distributor.

This fixes tpo/anti-censorship/rdsys#18