Split HTTPS distributor like Salmon.
According to our layers of abstraction, the HTTPS distributor consists
of a usecase and a presentation element.  The presentation element (the
actual Web server which accepts user requests) talks to the usecase (the
distributor logic).  We already implemented this separation for Salmon
and this patch now does it for HTTPS too.