Deduplicate the Web server code of our frontends.
Both the Salmon and HTTPS distributor frontends implement a Web API and
start their Web server in the exact same way.  This patch deduplicates
this code by turning it into a separate function.

In addition to the above, this patch simplifies our config code by
adding a dedicated WebApi object.