Add more information to a resource's state.
So far, we only saved a resource's state, i.e. if it's untested,
functional, or dysfunctional.  This patch makes rdsys also save when it
was last tested (according to bridgestrap) and what error occurred
during testing (if any).  This information is critical to bridge
operators who want to learn their bridge's status.

This patch partially fixes tpo/anti-censorship/rdsys#10