Refactor API of bridgestrap module.
So far, if rdsys was unable to talk to bridgestrap, it would accumulate
goroutines for pending test connections until rdsys eventually failed
due to an out of memory exception.

This patch fixes the problem by refactoring the API that talks to
bridgestrap and by not adding resources to the test pool if they're
already being tested.

This fixes tpo/anti-censorship/rdsys#30.